flask-backend/pyproject.toml
[project]
name = "flask-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.85",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "matplotlib>=3.10.1",
    "numpy>=2.2.5",
    "scikit-learn>=1.6.1",
]
